% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wnba_player_stats.R
\name{get_wnba_player_stats}
\alias{get_wnba_player_stats}
\title{Get 2024 Player Data from WNBA Statistics}
\usage{
get_wnba_player_stats(player_name)
}
\arguments{
\item{player_name}{A character string representing the name of the player whose data is to be retrieved.}
}
\value{
A \link[dplyr:reexports]{dplyr::tibble} containing the player's statistics if the player is found.
If the player is not found, a message indicating that the player was not found is returned.
}
\description{
This function retrieves the statistics of a specified WNBA player from the dataset.
The player's name is case-insensitive and is converted to uppercase for matching.
}
\details{
The function checks for the presence of the player in the dataset.
If exactly one match is found, it returns a summary of the player's statistics,
including points, rebounds, assists, and more. Additionally, it provides a link to the player's stats page.
}
\examples{
get_wnba_player_stats("Caitlin Clark")
get_wnba_player_stats("A'ja Wilson")
get_wnba_player_stats("Unknown Player")  # Should return a not found message

}
\seealso{
\link{wnba_stats_24}
}
\author{
Sophia Hillard
}
