% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_player_data.R
\docType{data}
\name{wnba}
\alias{wnba}
\title{WNBA Player Data
Player data from wnba repo}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3883 rows and 28 columns.
}
\usage{
wnba
}
\description{
WNBA Player Data
Player data from wnba repo
}
\keyword{datasets}
