% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_stat_grapher.R
\name{wnba_stat_grapher}
\alias{wnba_stat_grapher}
\title{This function graphs a statistic for one or more WNBA players selected from the data}
\usage{
wnba_stat_grapher(player_names, stat_type)
}
\arguments{
\item{player_names}{A character string; the name of the player(s) whose statistic will be graphed}

\item{stat_type}{A character string; the name of the basketball statistic that will be graphed}
}
\value{
A ggplot object: a graph showing the player's statistic over the time they've been in the league
(only if the player is in the data).
If the player isn't found, a message will appear to say that.
}
\description{
This function graphs a statistic for one or more WNBA players selected from the data
}
\details{
This function checks to see if the statistic type and player are in the data
if matches are found for both, it returns a graph of the player(s)' statistic over time.
This graph can be a comparison for multiple players or just show one player
}
\examples{
wnba_stat_grapher("Sue Bird", "f_tr")
wnba_stat_grapher(c("Sue Bird", "Candace Parker"), "f_tr")
wnba_stat_grapher("A'ja Wilson", "mp_pct") 

}
