% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_player_data.R, R/wnba_stats_24.R
\docType{data}
\name{wnba_stats_24}
\alias{wnba_stats_24}
\title{WNBA Player Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 157 rows and 30 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 157 rows and 30 columns.
}
\usage{
wnba_stats_24

wnba_stats_24
}
\description{
Player data from the 2024 WNBA season
}
\keyword{datasets}
