% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wnba}
\alias{wnba}
\title{WNBA Player Data}
\format{
a data frame with 28 variables and 3883 observations
}
\source{
FiveThirtyEight github repository \url{https://github.com/fivethirtyeight/WNBA-stats}
}
\usage{
wnba
}
\description{
WNBA Player Data
}
\keyword{datasets}
